Creating an IProduct interface for the array of products
Since Typescript is strongly typed it is a good practice to not leave any data type on 'any'. In a previous push I had an array of products with type any. For the current push I have created a product.ts file, which is an interface holding the properties for the description of the products. It helps showing error messages in the development phase when typing the properties for the product, showing error messages if some of the properties are spelled incorrectly. A huge benefit of having an interface is having strong types for the properties.
As a bonus, I have created a css file holding the future style for the elements in the product-list template. The only style change so far is the blue color of the thead element of the product table.