I have created a StarRating Component so that the numerical rating is displayed as number of stars
In the current push, I have created a nested StarComponent that changes the numerical value of the rating property of the products to actual star icons based on the number of the rating. I have experimented with @Input() and @Output() property decorators. The @Input() decorator helped me link the rating numbers from the parent component to  the nested (child) component and with a small mathematical function to determain how many stars it should have. The @Output() decorator helped me to set an EventEmitter for when a user clicks on the star rating of a product and then to be displayed in the parent component.