Moved the hardcoded products from the product component to a product service
Moved the hardcoded products from the product component to a product service.  In that case if I decide to get the products from an API, I am going to have to create a service to get the information. Using a service helps us with encapsulating parts of the code that is not related to a specific component and for reusability purpose.