ProductDetailComponent and WelcomeComponent
For this push I have created 2 new pages for the product app. The first one is the view for the product detail view. When the user clicks on the product he is going to be redirected to the view for the chosen product.
A WelcomeComponent is going to be the first page loaded when the web shop app is visited.
The next branch will be called RoutingConfiguration and as the name suggest the routing for the application will take place. I am going to configure the routing for the different views. Rewriting of the different URLs is going to take place.